+++ angular-src/src/app/services/muscle.service.ts
@@ -12,27 +12,22 @@ export class MuscleService {
 
 	getAll(){
 		return this.apiService.get(this.endPoint)
-      // .map(res => res.json())
 	}
 
 	get(id){
 		return this.apiService.get(this.endPoint+id)
-      // .map(res => res.json())
 	}
 
 	add(vo){
 		return this.apiService.post(this.endPoint, vo)
-      // .map(res => res.json())
 	}
 
 	delete(id){
 		return this.apiService.delete(this.endPoint+id)
-      // .map(res => res.json())
 	}
 
 	update(vo){
 		return this.apiService.put(this.endPoint+vo._id, vo)
-      // .map(res => res.json())
 	}
 
 }