+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## [2.0.0] - 2018-10-14
+### Added
+- CI with heroku
+- Load server env during initialization
+### Changed
+- update ApiService as a HTTP Factory 
+- migrated to github
+
 ## [1.1.1]
 ### Added
 - Update each exercise